+++ app/policies/event_policy.rb
@@ -1,7 +1,7 @@
 # Event policies
 class EventPolicy < ApplicationPolicy
   def show?
-    Event.available_for_user(user).pluck(:id).include?(record.id)
+    record.available_for_user?(user)
   end
 
   def edit?