+++ frontend/components_react/static_form/static_form.jsx
@@ -69,6 +69,8 @@ export default class StaticForm extends React.Component {
       url: url,
       data: { static: { name: state.name, staticable_type: staticableType, staticable_id: staticableId, description: state.description, privy: state.privy } },
       success: (data) => {
+        console.log(data)
+        console.log(data['static'])
         if (data['static'].guild_slug === null) window.location.replace(`${this.props.locale === 'en' ? '' : ('/' + this.props.locale)}/statics`)
         else window.location.replace(`${this.props.locale === 'en' ? '' : ('/' + this.props.locale)}/guilds/${data['static'].guild_slug}/management`)
       },
@@ -87,6 +89,8 @@ export default class StaticForm extends React.Component {
       url: url,
       data: { static: { name: state.name, description: state.description, privy: state.privy } },
       success: (data) => {
+        console.log(data)
+        console.log(data['static'])
         if (data['static'].guild_slug === null) window.location.replace(`${this.props.locale === 'en' ? '' : ('/' + this.props.locale)}/statics`)
         else window.location.replace(`${this.props.locale === 'en' ? '' : ('/' + this.props.locale)}/guilds/${data['static'].guild_slug}/management`)
       },